Interactions log and analysis - Validated
- Interactions log offers simple implementation of on-device storage and analysis
- Records encounters (time, proximity, payload) and includes function for enforcing retention period
- Analysis functions for reducing encounters by time window, payload, or proximity
- Provides basis for implementing on-device matching for contract tracing app
- Combine with payload matcher to filter encounters, then use analysis functions to compute basis for risk score estimation
