+++ herald/herald/Sensor/SensorArray.swift
@@ -26,7 +26,7 @@ public class SensorArray : NSObject, Sensor {
         super.init()
         
         // Loggers
-        add(delegate: EncounterLog(filename: "encounter.csv"))
+        add(delegate: Interactions(filename: "interactions.csv"))
         add(delegate: ContactLog(filename: "contacts.csv"))
         add(delegate: StatisticsLog(filename: "statistics.csv", payloadData: payloadData))
         add(delegate: DetectionLog(filename: "detection.csv", payloadData: payloadData))
